pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-eval-new"
version = "0.1.0"
description = "Azure AI Evaluation Framework for testing custom API agents - RAG quality and safety metrics"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv>=1.0.0",
    "azure-identity>=1.15.0",
    "azure-ai-evaluation>=1.0.0",
    "openai>=1.12.0",
    "pandas>=2.1.4",
    "jinja2>=3.1.2",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "flake8>=6.0.0",
    "pytest>=7.0.0",
    "pytest-mock>=3.12.0",
    "pre-commit>=3.0.0",
]

[tool.setuptools.packages.find]
include = ["agent_eval*"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
